columns + result grabbing abstracting
