+++ splinkclickhouse/clickhouse_server/dataframe.py
@@ -17,18 +17,9 @@ class ClickhouseServerDataFrame(ClickhouseDataFrame):
     def __init__(self, df_name, physical_name, db_api):
         super().__init__(df_name, physical_name, db_api)
 
-    @property
-    def columns(self) -> list[InputColumn]:
-        client = self.db_api.client
 
-        sql = self.db_api._information_schema_query(
-            "column_name", "columns", self.physical_name, self.db_api.database
-        )
 
-        res = client.query(sql).named_results()
-        cols = [r["column_name"] for r in res]
 
-        return [InputColumn(c, sqlglot_dialect_str="clickhouse") for c in cols]
 
     def validate(self):
         if not self.db_api.table_exists_in_database(self.physical_name):
